backend/requirements.txt
django>=4.0.0<=4.2.4
djangorestframework==3.14.0
pyyaml==6.0.1
requests==2.31.0
django-cors-headers==4.2.0
wheel
mysqlclient==2.1.1
